Tighten adapter test support (#114)
Reuse the adapter logger and simplify project-tag extraction.
Share the configurable metrics label collector across its tests.
